Moving functions not using REST calls to utils
