+++ lib/src/solid/api/rest_api.dart
@@ -32,21 +32,12 @@
 
 library;
 
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
-
 import 'package:fast_rsa/fast_rsa.dart';
 import 'package:http/http.dart' as http;
-import 'package:jwt_decoder/jwt_decoder.dart';
-import 'package:package_info_plus/package_info_plus.dart';
+import 'package:path/path.dart' as path;
 import 'package:rdflib/rdflib.dart';
 import 'package:solid_auth/solid_auth.dart';
-// ignore: implementation_imports
-import 'package:solid_auth/src/openid/openid_client.dart';
-import 'package:path/path.dart' as path;
 
-import 'package:solidpod/src/solid/common_func.dart';
 import 'package:solidpod/src/solid/constants.dart';
 import 'package:solidpod/src/solid/utils.dart';
 
@@ -328,66 +319,6 @@ Future<ResourceStatus> checkResourceExists(String resUrl, bool fileFlag) async {
   }
 }
 
-/// Generates a list of default folder paths for a given application.
-///
-/// This function takes the name of an application as input and returns a list of strings.
-/// Each string in the list represents a path to a default folder for the application.
-
-Future<List<String>> generateDefaultFolders() async {
-  final appName = await getAppName();
-  final mainResDir = appName;
-
-  final dataDirLoc = path.join(mainResDir, dataDir);
-  final sharingDirLoc = path.join(mainResDir, sharingDir);
-  final sharedDirLoc = path.join(mainResDir, sharedDir);
-  final encDirLoc = path.join(mainResDir, encDir);
-  final logDirLoc = path.join(mainResDir, logsDir);
-
-  final folders = [
-    mainResDir,
-    sharingDirLoc,
-    sharedDirLoc,
-    dataDirLoc,
-    encDirLoc,
-    logDirLoc,
-  ];
-  return folders;
-}
-
-/// Generates a list of default folder paths for a given application.
-///
-/// This function takes the name of an application as input and returns a list of strings.
-/// Each string in the list represents a path to a default folder for the application.
-
-Future<Map<dynamic, dynamic>> generateDefaultFiles() async {
-  final appName = await getAppName();
-  final mainResDir = appName;
-
-  const encKeyFile = 'enc-keys.ttl';
-  const pubKeyFile = 'public-key.ttl';
-  const indKeyFile = 'ind-keys.ttl';
-  const permLogFile = 'permissions-log.ttl';
-
-  final sharingDirLoc = path.join(mainResDir, sharingDir);
-  final sharedDirLoc = path.join(mainResDir, sharedDir);
-  final encDirLoc = path.join(mainResDir, encDir);
-  final logDirLoc = path.join(mainResDir, logsDir);
-
-  final files = {
-    sharingDirLoc: [
-      pubKeyFile,
-      '$pubKeyFile.acl',
-    ],
-    logDirLoc: [
-      permLogFile,
-      '$permLogFile.acl',
-    ],
-    sharedDirLoc: ['.acl'],
-    encDirLoc: [encKeyFile, indKeyFile],
-  };
-  return files;
-}
-
 /// Updates a file on the server with the provided SPARQL query.
 ///
 /// This asynchronous function sends a PATCH request to the server, targeting
@@ -645,244 +576,3 @@ Future<({String accessToken, String dPopToken})> getTokens(
 
 //   return fileUrl;
 // }
-
-/// Extract the app name and the version from the package info
-/// Return a record (with named fields https://dart.dev/language/records)
-
-Future<({String name, String version})> getAppNameVersion() async {
-  final info = await PackageInfo.fromPlatform();
-  return (name: info.appName, version: info.version);
-}
-
-/// Check whether a user is logged in or not
-///
-/// Check if the local storage has authentication
-/// details of the user and also check whether the
-/// access token is expired or not
-/// returns boolean
-
-Future<bool> checkLoggedIn() async {
-  final webId = await getWebId();
-
-  if (webId != null && webId.isNotEmpty) {
-    final accessToken = await AuthDataManager.getAccessToken();
-    if (accessToken != null && !JwtDecoder.isExpired(accessToken)) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-/// Delete login information from the local storage
-///
-/// returns true if successful
-
-Future<bool> deleteLogIn() async {
-  const success = true;
-  try {
-    await secureStorage.delete(key: 'webid');
-  } on Exception {
-    return false;
-  }
-  return success && (await AuthDataManager.removeAuthData());
-}
-
-/// Get the webId from local storage
-
-Future<String?> getWebId() async {
-  final webId = await secureStorage.read(key: 'webid');
-  return webId;
-}
-
-/// [AuthDataManager] is a class to manage auth data returned by
-/// solid-auth authenticate, including:
-/// - save auth data to secure storage
-/// - load auth data from secure storage
-/// - delete saved auth data from secure storage
-/// - refresh access token if necessary
-
-class AuthDataManager {
-  /// The URL for logging out
-  static String? _logoutUrl;
-
-  /// The RSA keypair and their JWK format
-  /// It seems Map<String, dynamic> does not work
-  static Map<dynamic, dynamic>? _rsaInfo;
-
-  /// The authentication response
-  static Credential? _authResponse;
-
-  /// The string key for storing auth data in secure storage
-  static const String _authDataSecureStorageKey = '_solid_auth_data';
-
-  /// Save the auth data returned by solid-auth authenticate in secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<void> saveAuthData(Map<dynamic, dynamic> authData) async {
-    const keys = [
-      'client',
-      'rsaInfo',
-      'authResponse',
-      'tokenResponse',
-      'accessToken',
-      'idToken',
-      'refreshToken',
-      'expiresIn',
-      'logoutUrl',
-    ];
-
-    for (final key in keys) {
-      assert(authData.containsKey(key));
-    }
-
-    _logoutUrl = authData['logoutUrl'] as String;
-    _rsaInfo = authData['rsaInfo'] as Map<dynamic,
-        dynamic>; // Note that use Map<String, dynamic> does not seem to work
-    _authResponse = authData['authResponse'] as Credential;
-
-    await writeToSecureStorage(
-        _authDataSecureStorageKey,
-        jsonEncode({
-          'logout_url': _logoutUrl,
-          'rsa_info': jsonEncode({
-            ..._rsaInfo!,
-            // Overwrite the 'rsa' keypair in rsaInfo
-            'rsa': {
-              'public_key': _rsaInfo!['rsa'].publicKey as String,
-              'private_key': _rsaInfo!['rsa'].privateKey as String,
-            },
-          }),
-          'auth_response': _authResponse!.toJson(),
-        }));
-
-    debugPrint('AuthDataManager => saveAuthData() done');
-  }
-
-  /// Retrieve (and reconstruct) auth data from secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<Map<dynamic, dynamic>?> loadAuthData() async {
-    if (_logoutUrl == null || _rsaInfo == null || _authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => loadAuthData() failed');
-        return null;
-      }
-    }
-
-    assert(_logoutUrl != null && _rsaInfo != null && _authResponse != null);
-    try {
-      final tokenResponse = await _getTokenResponse();
-      return {
-        'client': _authResponse!.client,
-        'rsaInfo': _rsaInfo,
-        'authResponse': _authResponse,
-        'tokenResponse': tokenResponse,
-        'accessToken': tokenResponse!.accessToken,
-        'idToken': _authResponse!.idToken,
-        'refreshToken': _authResponse!.refreshToken,
-        'expiresIn': tokenResponse.expiresIn,
-        'logoutUrl': _logoutUrl,
-      };
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => loadAuthData() failed: $e');
-    }
-    return null;
-  }
-
-  /// Remove/delete auth data from secure storage
-  static Future<bool> removeAuthData() async {
-    try {
-      await secureStorage.delete(key: _authDataSecureStorageKey);
-      _logoutUrl = null;
-      _rsaInfo = null;
-      _authResponse = null;
-
-      return true;
-    } on Exception {
-      debugPrint('AuthDataManager => removeAuthData() failed');
-    }
-    return false;
-  }
-
-  /// Returns the (refreshed) access token
-  static Future<String?> getAccessToken() async {
-    final tokenResponse = await _getTokenResponse();
-    if (tokenResponse != null) {
-      return tokenResponse.accessToken;
-    } else {
-      debugPrint('AuthDataManager => getAccessToken() failed');
-    }
-    return null;
-  }
-
-  /// Returns the (updated) token response
-  static Future<TokenResponse?> _getTokenResponse() async {
-    if (_authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => _getTokenResponse() failed');
-        return null;
-      }
-    }
-    assert(_authResponse != null);
-
-    try {
-      var tokenResponse = TokenResponse.fromJson(_authResponse!.response!);
-      if (JwtDecoder.isExpired(tokenResponse.accessToken!)) {
-        debugPrint(
-            'AuthDataManager => _getTokenResponse() refreshing expired token');
-        assert(_rsaInfo != null);
-        final rsaKeyPair = _rsaInfo!['rsa'] as KeyPair;
-        final publicKeyJwk = _rsaInfo!['pubKeyJwk'];
-        String tokenEndpoint =
-            _authResponse!.client.issuer.metadata['token_endpoint'] as String;
-        String dPopToken =
-            genDpopToken(tokenEndpoint, rsaKeyPair, publicKeyJwk, 'POST');
-        tokenResponse = await _authResponse!
-            .getTokenResponse(forceRefresh: true, dPoPToken: dPopToken);
-      }
-      return tokenResponse;
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => _getTokenResponse() failed: $e');
-    }
-    return null;
-  }
-
-  /// Returns the logout URL
-  static Future<String?> getLogoutUrl() async {
-    if (_logoutUrl == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => getLogoutUrl() failed');
-        return null;
-      }
-    }
-    assert(_logoutUrl != null);
-    return _logoutUrl;
-  }
-
-  /// Reconstruct the rsaInfo from JSON string
-  static Map<dynamic, dynamic> _getRsaInfo(String rsaJson) {
-    final rsaInfo_ = jsonDecode(rsaJson) as Map<String, dynamic>;
-    final publicKey = rsaInfo_['rsa']['public_key'] as String;
-    final privateKey = rsaInfo_['rsa']['private_key'] as String;
-
-    return {...rsaInfo_, 'rsa': KeyPair(publicKey, privateKey)};
-  }
-
-  /// Retrieve auth data from secure storage
-  static Future<bool> _loadData() async {
-    final dataStr = await secureStorage.read(key: _authDataSecureStorageKey);
-
-    if (dataStr != null) {
-      final dataMap = jsonDecode(dataStr) as Map<String, dynamic>;
-      _logoutUrl = dataMap['logout_url'] as String;
-      _rsaInfo = _getRsaInfo(dataMap['rsa_info'] as String);
-      _authResponse =
-          Credential.fromJson((dataMap['auth_response'] as Map).cast());
-
-      return true;
-    }
-    return false;
-  }
-}
